+++ src/main/java/io/github/zero88/jooq/vertx/datatype/JSONBToJsonPojoConverter.java
@@ -1,4 +1,4 @@
-package io.github.zero88.jooq.vertx.converter.ext;
+package io.github.zero88.jooq.vertx.datatype;
 
 import org.jooq.Converter;
 import org.jooq.JSONB;
