+++ src/main/java/io/github/zero88/jooq/vertx/datatype/PGJsonToVertxJsonBinding.java
@@ -1,4 +1,4 @@
-package io.github.zero88.jooq.vertx.converter.ext;
+package io.github.zero88.jooq.vertx.datatype;
 
 import java.sql.SQLException;
 import java.sql.SQLFeatureNotSupportedException;
